Better support for resourceType property in JSON
This change ensures that all resources have a resourceType. When going from JSON
into a go struct, the following will now happen:
* If a resourceType is provided and is correct, nothing will happen
* If it is not provided, it will be set in the struct
* If it is provided and wrong, an error will be returned

The same behavior is trur when going from a go struct into MongoDB

Also, the server will now return a 400 and OperationOutcome describing what went
wrong if a resource is PUT/POSTed with the wrong resourceType.